+++ src/render/glyph_atlas.ts
@@ -2,7 +2,7 @@ import {AlphaImage} from '../util/image.ts';
 import {register} from '../util/web_worker_transfer.ts';
 import potpack from 'potpack';
 
-import type {GlyphMetrics} from '../style/style_glyph.ts';
+import type {GlyphMetrics, GlyphVariants} from '../style/style_glyph.ts';
 import type {GetGlyphsResponse} from '../util/actor_messages.ts';
 
 const padding = 1;
@@ -28,32 +28,34 @@ export type GlyphPosition = {
 /**
  * The glyphs' positions
  */
-export type GlyphPositions = Record<string, Record<string, GlyphPosition>>;
+export type GlyphPositions = Record<string, GlyphVariants<Record<string, GlyphPosition>>>;
 
 export class GlyphAtlas {
     image: AlphaImage;
     positions: GlyphPositions;
 
     constructor(stacks: GetGlyphsResponse) {
-        const positions = {};
+        const positions: GlyphPositions = {};
         const bins = [];
 
         for (const stack in stacks) {
-            const glyphs = stacks[stack];
-            const stackPositions = positions[stack] = {};
+            const stackPositions = positions[stack] = {normal: {}, vertical: {}};
 
-            for (const id in glyphs) {
-                const src = glyphs[id];
-                if (!src || src.bitmap.width === 0 || src.bitmap.height === 0) continue;
+            for (const variant of ['normal', 'vertical'] as const) {
+                const glyphs = stacks[stack][variant];
+                for (const id in glyphs) {
+                    const src = glyphs[id];
+                    if (!src || src.bitmap.width === 0 || src.bitmap.height === 0) continue;
 
-                const bin = {
-                    x: 0,
-                    y: 0,
-                    w: src.bitmap.width + 2 * padding,
-                    h: src.bitmap.height + 2 * padding
-                };
-                bins.push(bin);
-                stackPositions[id] = {rect: bin, metrics: src.metrics};
+                    const bin = {
+                        x: 0,
+                        y: 0,
+                        w: src.bitmap.width + 2 * padding,
+                        h: src.bitmap.height + 2 * padding
+                    };
+                    bins.push(bin);
+                    stackPositions[variant][id] = {rect: bin, metrics: src.metrics};
+                }
             }
         }
 
@@ -61,13 +63,14 @@ export class GlyphAtlas {
         const image = new AlphaImage({width: w || 1, height: h || 1});
 
         for (const stack in stacks) {
-            const glyphs = stacks[stack];
-
-            for (const id in glyphs) {
-                const src = glyphs[id];
-                if (!src || src.bitmap.width === 0 || src.bitmap.height === 0) continue;
-                const bin = positions[stack][id].rect;
-                AlphaImage.copy(src.bitmap, image, {x: 0, y: 0}, {x: bin.x + padding, y: bin.y + padding}, src.bitmap);
+            for (const variant of ['normal', 'vertical'] as const) {
+                const glyphs = stacks[stack][variant];
+                for (const id in glyphs) {
+                    const src = glyphs[id];
+                    if (!src || src.bitmap.width === 0 || src.bitmap.height === 0) continue;
+                    const bin = positions[stack][variant][id].rect;
+                    AlphaImage.copy(src.bitmap, image, {x: 0, y: 0}, {x: bin.x + padding, y: bin.y + padding}, src.bitmap);
+                }
             }
         }